chore: restore text README
