+++ galactia/main.py
@@ -1,3 +1,4 @@
 from galactia.bot import run
+
 if __name__ == "__main__":
     run()
